refactor: extract capability::tier (the Tier enum)
Third step of the capability.rs module split, stacked on the
diagnostics.rs extraction. Pure lift-and-shift: the Tier enum and its
impl (plus tests) move into capability/tier.rs unchanged. capability.rs
re-exports Tier, since its still-inline guard and probe code depend on
it.

Depends on state.rs (already extracted); lands third.
